Slider vida funcional

+++ GameJam4/Assets/Prefabs/UI/HealthBar/HealthBar.cs
@@ -10,14 +10,14 @@ namespace UI
         public Slider slider;
         public Gradient gradient;
         public Image fill;
-        public void SetHealth(int health)
+        public void SetHealth(float health)
         {
             slider.value = health;
 
             fill.color = gradient.Evaluate(slider.normalizedValue);
         }
 
-        public void SetMaxHealth(int health)
+        public void SetMaxHealth(float health)
         {
             slider.maxValue = health;
             slider.value = health;
